Moving SamzaSqlRelRecord to samza-api as it is needed to be used in UDFs.
