+++ samza-sql/src/main/java/org/apache/samza/sql/serializers/SamzaSqlRelRecordSerdeFactory.java
@@ -23,37 +23,37 @@ import org.apache.samza.SamzaException;
 import org.apache.samza.config.Config;
 import org.apache.samza.serializers.Serde;
 import org.apache.samza.serializers.SerdeFactory;
-import org.apache.samza.sql.data.SamzaSqlRelMessage;
+import org.apache.samza.sql.SamzaSqlRelRecord;
 import org.codehaus.jackson.map.ObjectMapper;
 import org.codehaus.jackson.type.TypeReference;
 
 
 /**
- * A serializer for {@link SamzaSqlRelMessage.SamzaSqlRelRecord}. This serializer preserves the type information as
- * {@link SamzaSqlRelMessage.SamzaSqlRelRecord} and contains nested {@link SamzaSqlRelMessage.SamzaSqlRelRecord}
+ * A serializer for {@link SamzaSqlRelRecord}. This serializer preserves the type information as
+ * {@link SamzaSqlRelRecord} and contains nested {@link SamzaSqlRelRecord}
  * records.
  */
-public final class SamzaSqlRelRecordSerdeFactory implements SerdeFactory<SamzaSqlRelMessage.SamzaSqlRelRecord> {
-  public Serde<SamzaSqlRelMessage.SamzaSqlRelRecord> getSerde(String name, Config config) {
+public final class SamzaSqlRelRecordSerdeFactory implements SerdeFactory<SamzaSqlRelRecord> {
+  public Serde<SamzaSqlRelRecord> getSerde(String name, Config config) {
     return new SamzaSqlRelRecordSerde();
   }
 
-  public final static class SamzaSqlRelRecordSerde implements Serde<SamzaSqlRelMessage.SamzaSqlRelRecord> {
+  public final static class SamzaSqlRelRecordSerde implements Serde<SamzaSqlRelRecord> {
 
     @Override
-    public SamzaSqlRelMessage.SamzaSqlRelRecord fromBytes(byte[] bytes) {
+    public SamzaSqlRelRecord fromBytes(byte[] bytes) {
       try {
         ObjectMapper mapper = new ObjectMapper();
         // Enable object typing to handle nested records
         mapper.enableDefaultTyping();
-        return mapper.readValue(new String(bytes, "UTF-8"), new TypeReference<SamzaSqlRelMessage.SamzaSqlRelRecord>() {});
+        return mapper.readValue(new String(bytes, "UTF-8"), new TypeReference<SamzaSqlRelRecord>() {});
       } catch (Exception e) {
         throw new SamzaException(e);
       }
     }
 
     @Override
-    public byte[] toBytes(SamzaSqlRelMessage.SamzaSqlRelRecord p) {
+    public byte[] toBytes(SamzaSqlRelRecord p) {
       try {
         ObjectMapper mapper = new ObjectMapper();
         // Enable object typing to handle nested records
